avgProofTime uint16

+++ packages/protocol/contracts/L1/TaikoL1.sol
@@ -50,7 +50,7 @@ contract TaikoL1 is
         address _addressManager,
         bytes32 _genesisBlockHash,
         uint48 _initFeePerGas,
-        uint64 _initAvgProofWindow
+        uint16 _initAvgProofWindow
     )
         external
         initializer
